// src/app/shared/utils/enums/statuses.ts
export enum Statuses {
    ACTIVATED = "ACTIVATED",
    DEACTIVATED = "DEACTIVATED",
    AWAITING_PAYMENT = "AWAITING_PAYMENT",
    PENDING = "PENDING",
    IN_PROGRESS = "IN_PROGRESS",
    COMPLETED = "COMPLETED",
    DELIVERED = "DELIVERED",
    CANCELED = "CANCELED",
    OPENED = "OPENED",
    CLOSED = "CLOSED",
    DELETED = "DELETED",
}